+++ odroid/energymon-odroid.c
@@ -255,6 +255,7 @@ int energymon_init_odroid(energymon* em) {
   unsigned int i;
   char file[64];
   unsigned int count;
+  int err_save;
 
   // find the sensors
   char** sensor_dirs = get_sensor_directories(&count);
@@ -293,8 +294,10 @@ int energymon_init_odroid(energymon* em) {
     state->fds[i] = open(file, O_RDONLY);
     if (state->fds[i] < 0) {
       perror(file);
+      err_save = errno;
       free_sensor_directories(sensor_dirs, state->count);
       energymon_finish_odroid(em);
+      errno = err_save;
       return -1;
     }
   }
@@ -309,7 +312,9 @@ int energymon_init_odroid(energymon* em) {
   state->poll_sensors = 1;
   errno = pthread_create(&state->thread, NULL, odroid_poll_sensors, state);
   if (errno) {
+    err_save = errno;
     energymon_finish_odroid(em);
+    errno = err_save;
     return -1;
   }